submissions/api/views.py
from rest_framework import viewsets
from rest_framework.permissions import IsAuthenticated
from submissions.models import Submission
from .serializers import SubmissionSerializer


class SubmissionViewSet(viewsets.ModelViewSet):
    queryset = Submission.objects.all()
    serializer_class = SubmissionSerializer
    permission_classes = [IsAuthenticated]